Fix behaviour when communicating with some web browsers.
When communicating in SSL mode, we initially receive only the first character of a request from Chrome and Firefox. For example, for a GET request, we would only receive 'G'. The code used to assume the first line of the request would always arrive in one piece, and drop this request because it was invalid. These requests are now parsed and handled correctly.