Fix a crash in boost-asio when a connection closes while it is being handled.
If the connection closed, then the write of the response would notice this and
close the socket. However, if a read to the socket started at the same time,
there was a slight chance of TOCTTOU: at the start of the async_read the socket
would be open so the read would be started, but later the socket would be
invalid and its implementation 0x0, causing null pointer dereference.

A test is added that almost always failed before this patch, and now always
succeeds.